requirements.txt
django==4.0.0
numpy==1.21.5
python-stdnum==1.18
